add authentication information to the docs, add tags to openapi description to have proper menus in the docs, add description (as comments) to some endpoints which are rendered on the docs
