+++ api/src/lib.rs
@@ -2,6 +2,7 @@ pub mod auth;
 pub mod catalogs;
 pub mod conf;
 pub mod db;
+pub mod docs;
 pub mod filters;
 pub mod models;
 pub mod routes;
