+++ api/src/routes/mod.rs
@@ -1,6 +1,7 @@
 pub mod auth;
 pub mod catalogs;
 pub mod filters;
+pub mod info;
 pub mod queries;
 pub mod surveys;
 pub mod users;
